moves css into sibling folder
